requirements.txt
streamlit
pdfplumber
pandas
torch
sentence-transformers
pinecone-client
cohere
python-dotenv
